graph-datasources/graph-earnings-announcement-transcripts/load_data_gcs/requirements.txt
requests==2.31.0
python-dotenv==1.0.0
google-cloud-storage==2.14.0
